pyproject.toml
[project]
name = "text-annotation-preprocess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diffusers>=0.35.1",
    "google-auth>=2.40.3",
    "pydantic>=2.11.9",
    "setuptools>=80.9.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
]
